+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/doeverything/test/DoEverythingServerFeatureTest.scala
@@ -2133,4 +2133,79 @@ class DoEverythingServerFeatureTest extends FeatureTest {
     server.httpRequest(request, andExpect = Ok, withBody = s"$inMillis")
   }
 
+  test("/mustache.json") {
+    val response = server.httpRequest(
+      request = RequestBuilder.get("/mustache.json"),
+      andExpect = Ok,
+      withJsonBody = """{"name":"JSONDay"}"""
+    )
+
+    response.contentType should be(Some("application/json; charset=utf-8"))
+  }
+
+  test("/mustache-view-before.json") {
+    val response = server.httpRequest(
+      request = RequestBuilder.get("/mustache-view-before.json"),
+      andExpect = Ok,
+      withJsonBody = """{"name":"JSONDay"}"""
+    )
+
+    response.contentType should be(Some("application/json; charset=utf-8"))
+  }
+
+  test("/mustache-view-after.json") {
+    val response = server.httpRequest(
+      request = RequestBuilder.get("/mustache-view-before.json"),
+      andExpect = Ok,
+      withJsonBody = """{"name":"JSONDay"}"""
+    )
+
+    response.contentType should be(Some("application/json; charset=utf-8"))
+  }
+
+  test("/mustache-use-annotation-template-name.json") {
+    val response = server.httpRequest(
+      request = RequestBuilder.get("/mustache-use-annotation-template-name.json"),
+      andExpect = Ok,
+      withJsonBody = """{"name":"JSONDay"}"""
+    )
+
+    response.contentType should be(Some("application/json; charset=utf-8"))
+  }
+
+  test("/mustache-view-without-template-name-or-annotation.json") {
+    val response = server.httpRequest(
+      request = RequestBuilder.get("/mustache-view-without-template-name-or-annotation.json"),
+      andExpect = InternalServerError,
+      withJsonBody = """{"errors":["internal server error"]}"""
+    )
+
+    response.contentType should be(Some("application/json; charset=utf-8"))
+  }
+
+  test("/mustache.html") {
+    val response = server.httpRequest(
+      request = RequestBuilder.get("/mustache.html"),
+      andExpect = Ok,
+      withBody =
+        """<div class="nav">
+          |  <table cellpadding="0" cellspacing="0">
+          |    <tr>
+          |        <th>Name</th>
+          |        <th>Age</th>
+          |        <th>Friends</th>
+          |    </tr>
+          |    <tr>
+          |        <td>age2:42</td>
+          |        <td>name:HTMel</td>
+          |        <td>
+          |        </td>
+          |    </tr>
+          |  </table>
+          |</div>""".stripMargin
+    )
+
+    response.contentType should be(Some("text/html; charset=utf-8"))
+  }
+
 }